# backend/internal/httpapi/api.go
package httpapi

import (
	"context"
	"crypto/rand"
	"crypto/subtle"
	"encoding/hex"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"mime"
	"net"
	"net/http"
	"os"
	"runtime/debug"
	"strconv"
	"strings"
	"sync/atomic"
	"time"

	"gorm.io/gorm"

	"ianvs-terminal/backend/internal/auth"
	"ianvs-terminal/backend/internal/config"
	"ianvs-terminal/backend/internal/model"
	"ianvs-terminal/backend/internal/secure"
	"ianvs-terminal/backend/internal/store"
)

const (
	maximumBodySize      = 8 << 20
	maximumMergeBodySize = 32 << 20
	encryptionKeyHeader  = "X-Ianvs-Encryption-Key"
	requestIDHeader      = "X-Request-ID"
	maximumRequestIDSize = 128
)

var errUnsupportedMediaType = errors.New("content type must be application/json")
var requestIDFallback atomic.Uint64

type requestIDContextKey struct{}

type API struct {
	cfg    config.Config
	auth   *auth.Service
	store  *store.Store
	mux    *http.ServeMux
	logger *slog.Logger
}

type protectedHandler func(http.ResponseWriter, *http.Request, model.User, string)

func New(cfg config.Config, authService *auth.Service, resourceStore *store.Store) *API {
	return newWithLogger(
		cfg,
		authService,
		resourceStore,
		slog.New(slog.NewJSONHandler(os.Stderr, nil)),
	)
}

func newWithLogger(
	cfg config.Config,
	authService *auth.Service,
	resourceStore *store.Store,
	logger *slog.Logger,
) *API {
	api := &API{
		cfg:    cfg,
		auth:   authService,
		store:  resourceStore,
		mux:    http.NewServeMux(),
		logger: logger,
	}
	api.routes()
	return api
}

func (a *API) Handler() http.Handler {
	return a.withRecovery(a.withSecurityHeaders(a.withLocalBoundary(a.mux)))
}

func (a *API) routes() {
	a.mux.HandleFunc("GET /healthz", a.health)
	a.mux.HandleFunc("POST /v1/auth/setup", a.setupLocal)
	a.mux.HandleFunc("POST /v1/auth/register", a.register)
	a.mux.HandleFunc("POST /v1/auth/login", a.login)
	a.mux.Handle("POST /v1/auth/logout", a.protected(a.logout))
	a.mux.Handle("GET /v1/me", a.protected(a.me))
	a.mux.Handle("GET /v1/resources", a.protected(a.listResources))
	a.mux.Handle("GET /v1/resources/{kind}/{id}", a.protected(a.getResource))
	a.mux.Handle("PUT /v1/resources/{kind}/{id}", a.protected(a.putResource))
	a.mux.Handle("DELETE /v1/resources/{kind}/{id}", a.protected(a.deleteResource))
	a.mux.Handle("GET /v1/migrations/export", a.protected(a.exportMigration))
	a.mux.Handle("POST /v1/migrations/merge", a.protected(a.mergeMigration))
}

func (a *API) health(w http.ResponseWriter, r *http.Request) {
	var count int64
	if err := a.store.DB().WithContext(r.Context()).Model(&model.Setting{}).Count(&count).Error; err != nil {
		writeError(w, http.StatusServiceUnavailable, "database_unavailable", "database is unavailable")
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"status":    "ok",
		"mode":      a.cfg.Mode,
		"server_id": a.store.ServerID(),
		"time":      time.Now().UTC(),
	})
}

func (a *API) setupLocal(w http.ResponseWriter, r *http.Request) {
	if a.cfg.Mode != config.ModeLocal {
		writeError(w, http.StatusNotFound, "not_found", "endpoint is available only in local mode")
		return
	}
	if !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "encryption keys require HTTPS")
		return
	}
	var request struct {
		EncryptionKey string `json:"encryption_key"`
	}
	if err := decodeBody(w, r, maximumBodySize, &request); err != nil {
		writeDecodeError(w, err)
		return
	}
	user, created, err := a.auth.SetupLocalKey(r.Context(), request.EncryptionKey)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	status := http.StatusOK
	if created {
		status = http.StatusCreated
	}
	writeJSON(w, status, map[string]any{
		"user":        auth.View(user),
		"initialized": created,
	})
}

func (a *API) register(w http.ResponseWriter, r *http.Request) {
	if a.cfg.Mode != config.ModeRemote || !a.cfg.AllowRegistration {
		writeError(w, http.StatusNotFound, "not_found", "registration is not available")
		return
	}
	if !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "registration requires HTTPS")
		return
	}
	var request struct {
		Username      string `json:"username"`
		Password      string `json:"password"`
		EncryptionKey string `json:"encryption_key"`
	}
	if err := decodeBody(w, r, maximumBodySize, &request); err != nil {
		writeDecodeError(w, err)
		return
	}
	session, err := a.auth.Register(
		r.Context(),
		request.Username,
		request.Password,
		request.EncryptionKey,
	)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusCreated, session)
}

func (a *API) login(w http.ResponseWriter, r *http.Request) {
	if a.cfg.Mode != config.ModeRemote {
		writeError(w, http.StatusNotFound, "not_found", "login is not available in local mode")
		return
	}
	if !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "login requires HTTPS")
		return
	}
	var request struct {
		Username string `json:"username"`
		Password string `json:"password"`
	}
	if err := decodeBody(w, r, maximumBodySize, &request); err != nil {
		writeDecodeError(w, err)
		return
	}
	session, err := a.auth.Login(r.Context(), request.Username, request.Password)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, session)
}

func (a *API) logout(w http.ResponseWriter, r *http.Request, _ model.User, rawToken string) {
	if a.cfg.Mode == config.ModeRemote {
		if err := a.auth.Logout(r.Context(), rawToken); err != nil {
			a.writeServiceError(w, err)
			return
		}
	}
	w.WriteHeader(http.StatusNoContent)
}

func (a *API) me(w http.ResponseWriter, _ *http.Request, user model.User, _ string) {
	writeJSON(w, http.StatusOK, map[string]any{"user": auth.View(user)})
}

func (a *API) listResources(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	includeDeleted, err := queryBool(r, "include_deleted")
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_query", err.Error())
		return
	}
	includeSensitive, err := queryBool(r, "include_sensitive")
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_query", err.Error())
		return
	}
	key, err := a.keyForRequest(r, user, includeSensitive)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	resources, err := a.store.List(
		r.Context(),
		user,
		key,
		strings.TrimSpace(r.URL.Query().Get("kind")),
		includeDeleted,
		includeSensitive,
	)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{"resources": resources})
}

func (a *API) getResource(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	includeSensitive, err := queryBool(r, "include_sensitive")
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_query", err.Error())
		return
	}
	key, err := a.keyForRequest(r, user, includeSensitive)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	resource, err := a.store.Get(
		r.Context(),
		user,
		key,
		r.PathValue("kind"),
		r.PathValue("id"),
		includeSensitive,
	)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, resource)
}

func (a *API) putResource(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	var request struct {
		Data             json.RawMessage `json:"data"`
		Sensitive        json.RawMessage `json:"sensitive"`
		ClearSensitive   bool            `json:"clear_sensitive"`
		ExpectedRevision *int64          `json:"expected_revision"`
	}
	if err := decodeBody(w, r, maximumBodySize, &request); err != nil {
		writeDecodeError(w, err)
		return
	}
	sensitivePresent := len(request.Sensitive) > 0
	if sensitivePresent && request.ClearSensitive {
		writeError(w, http.StatusBadRequest, "invalid_request", "sensitive and clear_sensitive cannot be used together")
		return
	}
	if sensitivePresent && !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "sensitive data requires HTTPS")
		return
	}
	key, err := a.keyForRequest(r, user, sensitivePresent)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	resource, err := a.store.Put(
		r.Context(),
		user,
		key,
		r.PathValue("kind"),
		r.PathValue("id"),
		store.WriteInput{
			Data:             request.Data,
			Sensitive:        request.Sensitive,
			SensitivePresent: sensitivePresent,
			ClearSensitive:   request.ClearSensitive,
			ExpectedRevision: request.ExpectedRevision,
		},
	)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, resource)
}

func (a *API) deleteResource(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	var expectedRevision *int64
	if raw := strings.TrimSpace(r.URL.Query().Get("expected_revision")); raw != "" {
		value, err := strconv.ParseInt(raw, 10, 64)
		if err != nil || value <= 0 {
			writeError(w, http.StatusBadRequest, "invalid_query", "expected_revision must be a positive integer")
			return
		}
		expectedRevision = &value
	}
	if err := a.store.Delete(
		r.Context(),
		user,
		r.PathValue("kind"),
		r.PathValue("id"),
		expectedRevision,
	); err != nil {
		a.writeServiceError(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (a *API) exportMigration(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	includeDeleted, err := queryBool(r, "include_deleted")
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_query", err.Error())
		return
	}
	includeSensitive, err := queryBool(r, "include_sensitive")
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_query", err.Error())
		return
	}
	if includeSensitive && !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "sensitive exports require HTTPS")
		return
	}
	key, err := a.keyForRequest(r, user, includeSensitive)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	bundle, err := a.store.Export(r.Context(), user, key, includeDeleted, includeSensitive)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, bundle)
}

func (a *API) mergeMigration(w http.ResponseWriter, r *http.Request, user model.User, _ string) {
	var request store.MergeRequest
	if err := decodeBody(w, r, maximumMergeBodySize, &request); err != nil {
		writeDecodeError(w, err)
		return
	}
	needsKey := false
	for _, resource := range request.Resources {
		if len(resource.Sensitive) > 0 {
			needsKey = true
			break
		}
	}
	if needsKey && !a.sensitiveTransportAllowed(r) {
		writeError(w, http.StatusBadRequest, "secure_transport_required", "sensitive migrations require HTTPS")
		return
	}
	key, err := a.keyForRequest(r, user, needsKey)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	report, err := a.store.Merge(r.Context(), user, key, request)
	if err != nil {
		a.writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, report)
}

func (a *API) protected(next protectedHandler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if a.cfg.Mode == config.ModeLocal {
			user, err := a.auth.EnsureLocalUser(r.Context())
			if err != nil {
				a.writeServiceError(w, err)
				return
			}
			next(w, r, user, "")
			return
		}
		if !a.sensitiveTransportAllowed(r) {
			writeError(w, http.StatusBadRequest, "secure_transport_required", "remote API access requires HTTPS")
			return
		}
		rawToken := bearerToken(r.Header.Get("Authorization"))
		user, err := a.auth.AuthenticateToken(r.Context(), rawToken)
		if err != nil {
			writeError(w, http.StatusUnauthorized, "unauthorized", "a valid bearer token is required")
			return
		}
		next(w, r, user, rawToken)
	})
}

func (a *API) keyForRequest(r *http.Request, user model.User, required bool) ([]byte, error) {
	if !required {
		return nil, nil
	}
	return a.auth.VerifyKey(user, r.Header.Get(encryptionKeyHeader))
}

func (a *API) sensitiveTransportAllowed(r *http.Request) bool {
	if a.cfg.AllowInsecureSensitiveTransport || a.cfg.Mode == config.ModeLocal {
		return true
	}
	if r.TLS != nil {
		return true
	}
	return a.cfg.TrustProxyHeaders &&
		strings.EqualFold(strings.TrimSpace(r.Header.Get("X-Forwarded-Proto")), "https")
}

func (a *API) writeServiceError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, auth.ErrInvalidCredentials):
		writeError(w, http.StatusUnauthorized, "invalid_credentials", "username or password is invalid")
	case errors.Is(err, auth.ErrUsernameTaken):
		writeError(w, http.StatusConflict, "username_taken", err.Error())
	case errors.Is(err, auth.ErrInvalidUsername), errors.Is(err, auth.ErrInvalidPassword):
		writeError(w, http.StatusBadRequest, "invalid_account", err.Error())
	case errors.Is(err, secure.ErrKeyRequired):
		writeError(w, http.StatusPreconditionRequired, "encryption_key_required", "configure and provide the user encryption key")
	case errors.Is(err, secure.ErrInvalidKey):
		writeError(w, http.StatusUnauthorized, "invalid_encryption_key", "the encryption key is invalid")
	case errors.Is(err, secure.ErrWeakKey):
		writeError(w, http.StatusBadRequest, "weak_encryption_key", err.Error())
	case errors.Is(err, store.ErrNotFound), errors.Is(err, gorm.ErrRecordNotFound):
		writeError(w, http.StatusNotFound, "not_found", "resource was not found")
	case errors.Is(err, store.ErrRevisionConflict):
		writeError(w, http.StatusConflict, "revision_conflict", "resource revision changed")
	case errors.Is(err, store.ErrInvalidResource):
		writeError(w, http.StatusBadRequest, "invalid_resource", err.Error())
	default:
		writeError(w, http.StatusInternalServerError, "internal_error", "the request could not be completed")
	}
}

func (a *API) withLocalBoundary(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if a.cfg.Mode == config.ModeLocal {
			if !isLoopbackRequest(r) {
				writeError(w, http.StatusForbidden, "local_only", "local mode accepts loopback requests only")
				return
			}
			if a.cfg.LocalAccessToken != "" && !constantTimeEqual(
				bearerToken(r.Header.Get("Authorization")),
				a.cfg.LocalAccessToken,
			) {
				writeError(w, http.StatusUnauthorized, "local_access_denied", "a valid local access token is required")
				return
			}
		}
		next.ServeHTTP(w, r)
	})
}

func constantTimeEqual(actual, expected string) bool {
	if len(actual) != len(expected) {
		return false
	}
	return subtle.ConstantTimeCompare([]byte(actual), []byte(expected)) == 1
}

func (a *API) withSecurityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Cache-Control", "no-store")
		w.Header().Set("X-Content-Type-Options", "nosniff")
		w.Header().Set("X-Frame-Options", "DENY")
		next.ServeHTTP(w, r)
	})
}

func (a *API) withRecovery(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		requestID := validRequestID(r.Header.Get(requestIDHeader))
		if requestID == "" {
			requestID = newRequestID()
		}
		w.Header().Set(requestIDHeader, requestID)
		r = r.WithContext(context.WithValue(r.Context(), requestIDContextKey{}, requestID))
		response := &observedResponseWriter{
			ResponseWriter: w,
			status:         http.StatusOK,
		}
		started := time.Now()
		defer func() {
			if recovered := recover(); recovered != nil {
				if !response.wroteHeader {
					writeError(response, http.StatusInternalServerError, "internal_error", "the request could not be completed")
				}
				a.logRequestFailure(
					r,
					"http request panicked",
					response.status,
					started,
					"panic_type", fmt.Sprintf("%T", recovered),
					"stack", string(debug.Stack()),
				)
				return
			}
			if response.status >= http.StatusInternalServerError {
				a.logRequestFailure(
					r,
					"http request failed",
					response.status,
					started,
				)
			}
		}()
		next.ServeHTTP(response, r)
	})
}

type observedResponseWriter struct {
	http.ResponseWriter
	status      int
	wroteHeader bool
}

func (w *observedResponseWriter) WriteHeader(status int) {
	if w.wroteHeader {
		return
	}
	w.status = status
	w.wroteHeader = true
	w.ResponseWriter.WriteHeader(status)
}

func (w *observedResponseWriter) Write(body []byte) (int, error) {
	if !w.wroteHeader {
		w.WriteHeader(http.StatusOK)
	}
	return w.ResponseWriter.Write(body)
}

func (w *observedResponseWriter) Unwrap() http.ResponseWriter {
	return w.ResponseWriter
}

func (a *API) logRequestFailure(
	r *http.Request,
	message string,
	status int,
	started time.Time,
	extra ...any,
) {
	fields := []any{
		"request_id", requestIDFromContext(r.Context()),
		"method", r.Method,
		"route", requestRoute(r),
		"status", status,
		"duration_ms", time.Since(started).Milliseconds(),
	}
	fields = append(fields, extra...)
	a.logger.ErrorContext(r.Context(), message, fields...)
}

func requestRoute(r *http.Request) string {
	if r.Pattern == "" {
		return "unmatched"
	}
	return r.Pattern
}

func requestIDFromContext(ctx context.Context) string {
	requestID, _ := ctx.Value(requestIDContextKey{}).(string)
	return requestID
}

func validRequestID(raw string) string {
	requestID := strings.TrimSpace(raw)
	if requestID == "" || len(requestID) > maximumRequestIDSize {
		return ""
	}
	for _, character := range requestID {
		if (character >= 'a' && character <= 'z') ||
			(character >= 'A' && character <= 'Z') ||
			(character >= '0' && character <= '9') ||
			strings.ContainsRune("._:-", character) {
			continue
		}
		return ""
	}
	return requestID
}

func newRequestID() string {
	var random [16]byte
	if _, err := rand.Read(random[:]); err == nil {
		return hex.EncodeToString(random[:])
	}
	return fmt.Sprintf(
		"fallback-%d-%d",
		time.Now().UnixNano(),
		requestIDFallback.Add(1),
	)
}

func decodeBody(w http.ResponseWriter, r *http.Request, maximum int64, target any) error {
	mediaType, _, err := mime.ParseMediaType(r.Header.Get("Content-Type"))
	if err != nil || !strings.EqualFold(mediaType, "application/json") {
		return errUnsupportedMediaType
	}
	r.Body = http.MaxBytesReader(w, r.Body, maximum)
	decoder := json.NewDecoder(r.Body)
	decoder.DisallowUnknownFields()
	if err := decoder.Decode(target); err != nil {
		return err
	}
	var trailing any
	if err := decoder.Decode(&trailing); !errors.Is(err, io.EOF) {
		if err == nil {
			return errors.New("request body must contain one JSON value")
		}
		return err
	}
	return nil
}

func writeDecodeError(w http.ResponseWriter, err error) {
	if errors.Is(err, errUnsupportedMediaType) {
		writeError(w, http.StatusUnsupportedMediaType, "unsupported_media_type", err.Error())
		return
	}
	var maxBytesErr *http.MaxBytesError
	if errors.As(err, &maxBytesErr) {
		writeError(w, http.StatusRequestEntityTooLarge, "request_too_large", "request body is too large")
		return
	}
	writeError(w, http.StatusBadRequest, "invalid_json", "request body is not valid: "+err.Error())
}

func queryBool(r *http.Request, name string) (bool, error) {
	raw := strings.TrimSpace(r.URL.Query().Get(name))
	if raw == "" {
		return false, nil
	}
	value, err := strconv.ParseBool(raw)
	if err != nil {
		return false, fmt.Errorf("%s must be a boolean", name)
	}
	return value, nil
}

func bearerToken(header string) string {
	parts := strings.Fields(header)
	if len(parts) != 2 || !strings.EqualFold(parts[0], "Bearer") {
		return ""
	}
	return parts[1]
}

func isLoopbackRequest(r *http.Request) bool {
	host, _, err := net.SplitHostPort(r.RemoteAddr)
	if err != nil {
		host = r.RemoteAddr
	}
	address := net.ParseIP(strings.Trim(host, "[]"))
	return address != nil && address.IsLoopback()
}

func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}

func writeError(w http.ResponseWriter, status int, code, message string) {
	writeJSON(w, status, map[string]any{
		"error": map[string]string{
			"code":    code,
			"message": message,
		},
	})
}

// ShutdownContext provides a bounded context shared by the command entrypoint.
func ShutdownContext(parent context.Context) (context.Context, context.CancelFunc) {
	return context.WithTimeout(parent, 10*time.Second)
}